+++ src/api/mod.rs
@@ -1,0 +1,9 @@
+mod routes;
+pub mod server;
+
+pub enum HTTPError {
+    NotFound,
+    BadRequest,
+    NotAuthorized,
+    ServerError,
+}
